app/retrieval/requirements.txt
openai-agents
python-dotenv
psycopg[binary]
requests
fastapi
uvicorn
feedparser